implement subscription `posixTimeIsPast` in platform web service
Expand the JS-free Elm app execution path to support the subscription `posixTimeIsPast` of the web service platform: Fix parsing of this subscription and add interface methods.

+++ implement/Pine.Core/ElmInteractive/ElmValueEncoding.cs
@@ -535,7 +535,7 @@ public static class ElmValueEncoding
                 .Extract(err => throw new Exception(err)),
 
                 ElmValue.ElmInteger elmInteger =>
-                PineValueAsInteger.ValueFromSignedInteger((int)elmInteger.Value),
+                PineValueAsInteger.ValueFromSignedInteger(elmInteger.Value),
 
                 ElmValue.ElmString elmString =>
                 StringAsPineValue(elmString.Value),
